soc/amd/psp_smm_gen2.c: Fix 64bit mode integer conversion
Explicitly cast integers to fix building for long mode.

Reviewed-on: https://review.coreboot.org/c/coreboot/+/69507
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>

+++ src/soc/amd/common/block/psp/psp_smm_gen2.c
@@ -4,6 +4,7 @@
 #include <amdblocks/acpimmio.h>
 #include <amdblocks/psp.h>
 #include <amdblocks/smi.h>
+#include <stdint.h>
 
 void soc_fill_smm_trig_info(struct smm_trigger_info *trig)
 {
@@ -13,8 +14,8 @@ void soc_fill_smm_trig_info(struct smm_trigger_info *trig)
 	trig->address = (uintptr_t)acpimmio_smi + SMI_REG_SMITRIG0;
 	trig->address_type = SMM_TRIGGER_MEM;
 	trig->value_width = SMM_TRIGGER_DWORD;
-	trig->value_and_mask = ~SMITRIG0_PSP;
-	trig->value_or_mask = SMITRIG0_PSP;
+	trig->value_and_mask = (uint32_t)~SMITRIG0_PSP;
+	trig->value_or_mask = (uint32_t)SMITRIG0_PSP;
 }
 
 void soc_fill_smm_reg_info(struct smm_register_info *reg)
